chore(deps): break utils/techBaseValidation 3-way cycle via leaf types module
Extracts ComponentSelections + ComponentValidator + COMPONENT_VALIDATORS
(plus the createValidator factory) into a new leaf module
techBaseValidation.types.ts. The orchestrator (techBaseValidation.ts),
the helpers (techBaseValidation.helpers.ts), and the affected-selection
mapping (techBaseValidation.mapping.ts) now all import shared symbols
from the leaf rather than from each other.

Cycle closed:
  techBaseValidation.helpers.ts > techBaseValidation.mapping.ts >
  techBaseValidation.ts (which previously re-imported helpers)

The orchestrator continues to re-export ComponentSelections,
ComponentValidator, and COMPONENT_VALIDATORS for backward compatibility,
so all existing consumers (e.g. useUnitTechBaseStore) work unchanged.

The new leaf depends only on techBaseValidation.filters.ts, which is
already an independent leaf.

No behavior change. Public API surface (exports from
@/utils/techBaseValidation) is identical.
